Add terminal functionality
Client can now type their bid in the terminal and press enter to send it.
The program waits for an input, then sends it through RPC.
Afterwards, the result function is called which notifies the client
about the highest bid.

Also, PlaceBid() and Result() are in an infinite loop now, so the client
can bid however much they want