Add node basic reasoning about bids
When a node receives bid, it now checks the current highest bid against the new
received bid. If the new bid is higher, the node sets it and the bidder ID as the highest,
and prints it in its own terminal.

Also, on a client calling Result(), it sends a string revealing the current highest bid
and the Id of the bidder.